Split out getEditedPostSlug
